requirements.txt
numpy~=2.4.0
requests~=2.32.5
python-dotenv~=1.2.1
groq~=1.0.0
protobuf~=5.29.5